move timetable files to /imports/
