+++ client/main.scss
@@ -42,6 +42,7 @@
 @import '{}/imports/ui/components/sharing/sharing';
 
 @import '{}/imports/ui/pages/calendar/calendar';
+@import '{}/imports/ui/pages/timetable/timetable';
 
 // Stylesheets from old structure outside of imports
 @import 'views/views';
