+++ client/views/_views.scss
@@ -57,9 +57,6 @@
 @import 'static/notFound/notFound';
 @import 'static/translate/translate.info';
 
-// Timetable
-@import 'timetable/timetable';
-
 // Users
 @import 'users/courselist/users.courselist';
 @import 'users/ownprofile/ownprofile';
