+++ imports/ui/components/navbar/navbar.js
@@ -1,7 +1,7 @@
 import { Meteor } from 'meteor/meteor';
 import { Router } from 'meteor/iron:router';
 import { Session} from 'meteor/session';
-import { Template } from 'meteor/tempating';
+import { Template } from 'meteor/templating';
 import { $ } from 'meteor/jquery';
 
 import ScssVars from '/imports/ui/lib/scss-vars.js';
